copied the rebased fork (#151)
* Changed uint usages to use ints instead wherever possible


* ran local tests so that json types are updated accordingly as well

+++ contracts/hts-precompile/ExpiryHelper.sol
@@ -8,13 +8,13 @@ abstract contract ExpiryHelper {
 
     function createAutoRenewExpiry(
         address autoRenewAccount,
-        uint32 autoRenewPeriod
+        int64 autoRenewPeriod
     ) internal pure returns (IHederaTokenService.Expiry memory expiry) {
         expiry.autoRenewAccount = autoRenewAccount;
         expiry.autoRenewPeriod = autoRenewPeriod;
     }
 
-    function createSecondExpiry(uint32 second) internal pure returns (IHederaTokenService.Expiry memory expiry) {
+    function createSecondExpiry(int64 second) internal pure returns (IHederaTokenService.Expiry memory expiry) {
         expiry.second = second;
     }
 }
