+++ contracts/hts-precompile/examples/token-manage/TokenManagementContract.sol
@@ -70,7 +70,7 @@ contract TokenManagementContract is HederaTokenService, ExpiryHelper, KeyHelper 
         emit UnpausedToken(true);
     }
 
-    function wipeTokenAccountPublic(address token, address account, uint32 amount) public returns (int responseCode) {
+    function wipeTokenAccountPublic(address token, address account, int64 amount) public returns (int responseCode) {
         responseCode = HederaTokenService.wipeTokenAccount(token, account, amount);
         emit ResponseCode(responseCode);
 
@@ -92,7 +92,7 @@ contract TokenManagementContract is HederaTokenService, ExpiryHelper, KeyHelper 
         (responseCode) = HederaTokenService.updateTokenInfo(token, tokenInfo);
         emit ResponseCode(responseCode);
 
-        if(responseCode != HederaResponseCodes.SUCCESS) {
+        if (responseCode != HederaResponseCodes.SUCCESS) {
             revert();
         }
     }
@@ -101,7 +101,7 @@ contract TokenManagementContract is HederaTokenService, ExpiryHelper, KeyHelper 
         (responseCode) = HederaTokenService.updateTokenExpiryInfo(token, expiryInfo);
         emit ResponseCode(responseCode);
 
-        if(responseCode != HederaResponseCodes.SUCCESS) {
+        if (responseCode != HederaResponseCodes.SUCCESS) {
             revert();
         }
     }
@@ -110,12 +110,11 @@ contract TokenManagementContract is HederaTokenService, ExpiryHelper, KeyHelper 
         (responseCode) = HederaTokenService.updateTokenKeys(token, keys);
         emit ResponseCode(responseCode);
 
-        if(responseCode != HederaResponseCodes.SUCCESS) {
+        if (responseCode != HederaResponseCodes.SUCCESS) {
             revert();
         }
     }
-
-    function burnTokenPublic(address token, uint64 amount, int64[] memory serialNumbers) external returns (int256 responseCode, uint64 newTotalSupply) {
+    function burnTokenPublic(address token, int64 amount, int64[] memory serialNumbers) external returns (int256 responseCode, int64 newTotalSupply) {
         (responseCode, newTotalSupply) = HederaTokenService.burnToken(token, amount, serialNumbers);
         emit ResponseCode(responseCode);
